Add non-deterministic workflow option (#475)
This commit adds an option to configure how worker execution should
handle detection of a non-deterministic workflow.

Currently, when non-determinism is detected in execution history, the
task handler on the client simply logs the error to console *without*
replying back to the cadence server, which leads the server to think
the decision task has timed out and thus reschedules a new decsion task.

The currently behavior is less than satisfactory for 2 reasons:
1. From server side there's no indication of the non-determinism error.
   It simply looks like a timeout.
2. From the client side, there's no recover option in that the client
   will just pull off the new rescheduled decision task and fail again
   and again.

This commit adds a new option to cadence workers to explicit fail the
workflow where a mismatched history has been detected. This should
yield a more robust debugging experience for customers whose workflows
fail due to non-determinism.